+++ apps/remnant2toolkit/app/(enemies)/world-save-archive/_types/index.ts
@@ -3,16 +3,16 @@ import { type ReleaseKey } from '@/app/_types/releases';
 import { type ALL_BOSS_AFFIXES } from '@/app/(enemies)/world-save-archive/_constants/all-boss-affixes';
 import { type ALL_BOSSES } from '@/app/(enemies)/world-save-archive/_constants/all-bosses';
 
+export type BossAffix = keyof typeof ALL_BOSS_AFFIXES;
+export type BossAffixName = (typeof ALL_BOSS_AFFIXES)[number]['name'];
+
+export type BossName = (typeof ALL_BOSSES)[number]['name'];
+
 export interface WorldSaveFilters {
   bossName: string | DefaultFilter;
   bossAffixes: string[] | [DefaultFilter];
   releases: string[] | [DefaultFilter];
 }
-
-export type BossAffix = keyof typeof ALL_BOSS_AFFIXES;
-export type BossAffixName = (typeof ALL_BOSS_AFFIXES)[number]['name'];
-
-export type BossName = (typeof ALL_BOSSES)[number]['name'];
 
 export interface WorldSave {
   bossName: BossName;